+++ src/test/java/teammates/test/pageobjects/InstructorFeedbackSessionsPage.java
@@ -216,7 +216,7 @@ public class InstructorFeedbackSessionsPage extends AppPage {
             LocalDateTime visibleTime,
             LocalDateTime publishTime,
             Text instructions,
-            int gracePeriod) {
+            long gracePeriod) {
 
         fillTextBox(fsNameTextBox, feedbackSessionName);
 
@@ -236,7 +236,7 @@ public class InstructorFeedbackSessionsPage extends AppPage {
 
         // Select grace period
         if (gracePeriod != -1) {
-            selectDropdownByVisibleValue(gracePeriodDropdown, Integer.toString(gracePeriod) + " mins");
+            selectDropdownByVisibleValue(gracePeriodDropdown, Long.toString(gracePeriod) + " mins");
         }
 
         clickSubmitButton();
@@ -244,7 +244,7 @@ public class InstructorFeedbackSessionsPage extends AppPage {
 
     public void addFeedbackSessionWithTimeZone(String feedbackSessionName, String courseId,
             LocalDateTime startTime, LocalDateTime endTime, LocalDateTime visibleTime, LocalDateTime publishTime,
-            Text instructions, int gracePeriod, ZoneId timeZone) {
+            Text instructions, long gracePeriod, ZoneId timeZone) {
 
         selectTimeZone(timeZone);
 
@@ -254,7 +254,7 @@ public class InstructorFeedbackSessionsPage extends AppPage {
 
     public void addFeedbackSessionWithStandardTimeZone(String feedbackSessionName, String courseId,
             LocalDateTime startTime, LocalDateTime endTime, LocalDateTime visibleTime, LocalDateTime publishTime,
-            Text instructions, int gracePeriod) {
+            Text instructions, long gracePeriod) {
 
         addFeedbackSessionWithTimeZone(feedbackSessionName, courseId, startTime, endTime, visibleTime, publishTime,
                 instructions, gracePeriod, ZoneId.of("UTC+08:00"));
